Fix decoupled sink semantics
NASTI doesn't seem to handle streaming requests to different slaves
gracefully (AW requests need to be buffered; the route cannot change
until the W req is completed, but instead is reassigned immediately
on an AW handshake). This change allows the the tester to more carefully
monitor the number of outstanding requests.