[MIDAS] Permit FAME1-transformations of modules within a hierarchy.
Previously, our F1 transformation applied only to an entire circuit --
we BBoxed in the transformed RTL into a larger host-circuit.

This change permits instances within a module hierachy to be
transformed. In chisel, the FAME1-root instance is annotated with F1 and
no-dedup annotations. The annotation also contains the name of the
host-signal that will be used as targetFire. In this pass:

1) For each F1-root, all modules in the root's child-instance hierarchy are
duplicated and renamed with an _f1 suffix. Module definitions that are not
instantiated in a non-F1(host) context are removed.

2) All F1-module definitions are transfomed using the standard pass.

3) F1-root instances have their targetFire ports connected to the annotated
targetFire signal in their instantiating host module.