New squashed FireSim repo for open-sourcing
This is the new FireSim repo that will be open-sourced. This commit
matches the state of the repo at commit
031132485840608f111e6c745d8140b5def6d857 of the original FireSim
repo, which now lives at
https://github.com/firesim/firesim-prerelease

All future FireSim development will happen in this repository. The
FireSim pre-release repo is maintained privately only for archival
purposes (e.g. if someone needs to do git blame on a file) - no new
commits should be made on the pre-release repo.