[Driver] Make read_chunk bypass to loadmem during program load
For programs that were not aligned to 32 bytes, fesvr would attempt
to read that chunk before writing out the last chunk of the program.
This read would not bypass to loadmem -- it would fetch a
line into cache -- however, the final write would bypass to DRAM, leaving
the line in  outer cache that would not be invalidated.