Allow annotations to flow through entire compiler
This was a more invasive change than i was hoping, as it required
updating names in both the outer and inner modules during linkage in
SimulationMapping and PlaformMapping. Adds a PreLinkRenaming
transforming which does the renaming for the outer module.

This is the basis for supporting custom transformations that
should be scheduled after the MIDAS transformations.

Finally, adds a FpgaDebug annotation that will be implemented by the
ILATopWiring Transformation, but in the future could be emitted
as (* mark_debug = "true" *) by a smarter emitter.