setup: Ignore SSH agent forwarding with tmux and screen
Ensure that a local ssh-agent instance is always used if running within
tmux or screen.  Previously, a separate instance would not be started
when agent forwarding is active.  However, detaching the session and
logging out would then close the agent socket, usually causing ongoing
manager jobs to fail when attempting to rsync results.