Enable customruntimeconfig to be relative path above firesim
When using FireSim as a library in another project, it is useful to keep
configuration files in the repo of the toplevel project. This is a minimal
change that enables a user to provide a path like ../../../toplevel-configs/somegreat.conf

infrasetup already copies the file correctly.  This change only modifies
RuntimeHWConfig.get_local_runtimeconf_binaryname() to explicitly strip the
path to os.path.basename