add pytest driven deploy/tests
Adding two tests for the new awstools.get_snsname_arn() function.

To run them:
* make sure you have the new deps in machine-launch-script.sh
* cd deploy
* pytest

Useful primers on pytest & testing AWS:
* https://tensoriot.medium.com/unit-testing-with-pytest-and-moto-e94fc2eefe7a
* https://github.com/boto/boto3/issues/2485

Useful primer on unittest.mock (fka py-mock, not to be confused with pymock):
* https://www.fugue.co/blog/2016-02-11-python-mocking-101

'mock' became part of the stdlib in python 3.3 and was backported to 2.7
as 'py-mock'.  Of course, python being python, there is also a pymock
and that's totally different.

Detailed walkthrough of credential protection while using moto:
* https://blog.codecentric.de/en/2020/01/testing-aws-python-code-with-moto/