refactor usage of StreamLogger
This commit puts call of firesim.main into a StreamLogger context so
that all output to stdout will be converted to DEBUG logs and all
output to stderr will be converted to INFO logs.

All other uses of StreamLogger are removed, except for when we
want stdout to become INFO logs.  Then, we retain the existing InfoStreamLogger
contexts.

This addresses cases where StreamLogger was missing around code
that explicitly printed to sys.stdout or sys.stderr and output
was failing to be logged.