rework buildafi.get_sim_makevar to work for regular makevars too
previous implementation using '--dry-run --print-data-base' only
worked for 'simple' (i.e. := ) makevars.  Regular (i.e. =) makevars
are not expanded in the data-base dump.  Instead, implemented a phony
'echo' target that takes a manditory ECHOVAR=<varname> parameter
and prints the value of that variable as a line:
<varname> = <value>

to stdout if it exists.  If it is undefined, it prints a message to stderr
and causes make to exit non-zero.