firesim tasks throw if not called via TASKS
When patching for testing, I was concerned about where to do the patching
and making the symbol table lookups throw seemed like a sensible thing.  However,
now that I'm done writing the tests, this seems like it might be overkill
and that's why I'm leaving it in a separate commit in case I want to revert it.