add test to demonstrate moto doesn't care about inst type
Abe was curious whether boto would error if you give it a bogus instance
type when launching an instance.  I added a test that shows what moto will
do.  The boto test might be a good example for how to un-munge the environment
for actually using AWS in unit tests that don't mock it out.