consistently use CXXFLAGS for FireSim C++
There was a mixture of using CXXFLAGS and CFLAGS in the FireSim driver makefiles
this led to needing to pass both of them to VCS and Verilator.  However, everything
we're building is C++ and so is Verilator, so we should only need to use CXXFLAGS.

This cuts down on the size of the compile commandlines and makes them much easier to reason about.