Removed the Fame1Instances transformation (#1202)
The `Fame1Instances` transformation was used solely to clock gate the model within `FasedMemoryTimingModel`. Instead of the transformation which individually altered all memories and registers within the hierarchy, now a clock gate is used on the clock signal passed to the model.

Altered `HostSpecialization` for Xilinx to replace all modules which have `AbstractClockGate` in their name since the elaboration of the simulation wrapper generates the `AbstractClockGate_0` name to avoid duplication.