Introduced an interface to capture the user-defined logic of a simulation (#1328)
Instead of chains of derivation, the user-defined part of a simulation now
must implement the `simulation_t` interface. The `simif_t` implementation
takes a reference to this interface and runs it.

These changes are useful as they define a uniform interface across tests
and the actual FireSim driver.

This PR also nukes the ucontext-based implementation from `context.h` and
renames the class into `pcontext` to avoid linking conflicts with the
same class from libfesvr.a.