Remove the use of DPI utilities from dpi.cc
dpi.cc accesses the underlying arrays using bit manipulation functions.
This removes the reliance on the utilities provided in `svdpi.h`, allowing
it to be independently compiled and fully shared with verilator.

This change also excludes the generated from C sources, as VCS attempted
to compile a file with it included.