Re-enable compilation of the Dromajo bridge (#1384)
The Dromajo bit suffers from a fair bit of bit rot. This patch fixes breakages from stream engines and MMIO port rewrites.
The Dromajo bridge is now compiled as part of the support library, ensuring at least the C++ side is kept updated.
Due to its use of an additional header, it is an interesting bridge to keep maintaining.
The bridge is not tested as the Scala counterpart does not compile with Chipyard at the moment.