Enable clang-tidy on all C++ sources (#1400)
This PR adds a new make target, `clang-tidy` to run clang-tidy on all C++ sources.
Only a dummy check is enabled, as this patch applies the minimal number of fixes
to compile headers and sources with clang without the header. In a subsequent PR,
more checks will be enabled and the files will be formatted.

`clang-tidy` can be executed using `make -C sim clang-tidy`. It will automatically
apply fixes to known issues.