Added a test for memory accesses
This PR fixes the invocation to `LoadMem` and adds a test for it.

The test performs memory read accesses through a TileLink node which goes to DRAM through a trivial FASED bridge.
Read requests from the driver are received through the `PeekPokeBridge` and are sent back to be printed to a file.
The test harness compares the file passed to loadmem with the results produced by this bridge, which should read all data back.